# services/reference-service/src/main/resources/openapi/reference-api.yaml
...
openapi: 3.1.0
info:
  title: CardDemo Reference Service API
  version: 1.0.0
  summary: >-
    Reference data of the migrated CardDemo credit-card management application:
    transaction types and their categories, disclosure-group interest rates, the
    three United States address lookups, and date evaluation.
  description: >-
    The reference context owns the data every other context reads and none of
    them writes. It publishes nineteen operations, across thirteen paths, over
    six tables.


    Transaction types and transaction categories are maintained here, and the
    relationship between them is the one referential rule the baseline states
    anywhere in this domain: a category references its type, and the type cannot
    be deleted while a category still references it. That refusal is a
    first-class user outcome in the baseline rather than a fault, and it is
    reported here as 409 for the reason recorded on the Conflict response.


    Disclosure-group rows carry the interest rate the nightly interest
    calculation consumes. That rate is an arithmetic operand, not a display
    value, so it is transported as a JSON string for the reason recorded on the
    InterestRate schema, and the lookup operation documents the DEFAULT-group
    fallback the baseline performs when an account's own group has no matching
    row.


    The three address lookups -- phone area codes, state codes and state and
...
    operation that writes them.


...


...
    one field.
...

...
servers:
...
security:
  - bearerAuth: []

# WHY : Refactoring Rationale: the authority each operation requires is declared
#       as a machine-readable field on the operation itself, and not only in
#       prose and tag membership. Prose and tag membership are not checkable: a
#       reviewer can read that a write is administrative, and a generator or a
#       gateway cannot, so the restriction would be indistinguishable from any
#       other authenticated route. Every operation below carries
#       x-required-authority.
# WHY : Assumptions: the value on each operation is not chosen here. It is
#       derived from what SecurityConfig.java already enforces for this service:
#       POST, PUT, PATCH and DELETE on every path each require
#       JwtRoleConverter.ADMIN_AUTHORITY, at its lines 99 to 106, and every
#       remaining request needs only authentication, at line 107. So every write
#       below carries carddemo-admin and every READ carries the third value
#       authenticated, and the document describes the filter chain rather than a
#       second opinion about it.
# WHY : Refactoring Rationale: an earlier revision carried carddemo-user on the
#       reads, which stated something the filter chain does not enforce and could
#       not be enforced without a behavioural regression. The shared converter
#       grants exactly the groups a token presents and builds NO hierarchy -- its
#       conversion loop admits a group only if it is one of the two recognised
#       names -- so a token carrying carddemo-admin alone receives ADMIN_AUTHORITY
#       and nothing else. A gateway or generator enforcing carddemo-user on the
#       reads would therefore refuse every administrator, and the baseline refuses
#       an administrator nothing: app/cbl/COSGN00C.cbl sends type 'A' to the
#       administrative menu at its line 245 and every function is reachable from
#       there. The honest declaration is the one the chain makes.
# WHY : Alternatives Considered: granting USER_AUTHORITY to an administrator as
#       well, so that carddemo-user could be declared and enforced on the reads.
#       Rejected because the grant would have to be made in the shared converter,
#       where it would change the authority set every one of the nine contexts
#       sees, to buy nothing this service needs: a read that any authenticated
#       caller may perform is described exactly by saying so. The two GROUP values
#       remain exactly the two the shared converter recognises, ADMIN_AUTHORITY
#       and USER_AUTHORITY at
#       services/common-lib/src/main/java/com/carddemo/common/security/JwtRoleConverter.java
#       lines 250 and 260, which are the migrated form of the two SEC-USR-TYPE
#       values 'A' and 'U'; authenticated names a REQUIREMENT rather than a group,
#       and the model below distinguishes the two.
...
paths:
  /api/v1/reference/transaction-types:

    description: >-
      The collection of transaction types, addressable as a key-ordered sequence
      of pages and as the place a new type is created. A type is the parent of
      the categories beneath it, which is why it is the first collection in this
      document and why its delete operation can be refused.

    get:
      tags:
        - reference-data
      operationId: listTransactionTypes
      x-required-authority: authenticated
      summary: List transaction types by key, one page at a time
      description: >-
        Returns one page of transaction types in ascending type-code order,
        optionally narrowed by a type-code filter, a description filter or both.
        This is the migrated form of the browse the baseline drives with a pair
        of declared cursors at app/app-transaction-type-db2/cbl/COTRTLIC.cbl
        lines 338 to 368, and it preserves those cursors' page boundaries
        exactly, including the difference in strictness between reading forward
        and reading backward that is set out on the cursor parameter.


        Paging is by key and never by ordinal. Send no cursor to obtain the
        opening page; thereafter copy lastKey from the page in hand and send
        direction next to advance, or copy firstKey and send direction previous
        to retreat. When the filters exclude every row the page is empty and
        both cursors are null, which the baseline reports with the message text
        recorded on the TransactionTypePage schema.
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PagingDirection'
        - $ref: '#/components/parameters/TypeCodeFilter'
        - $ref: '#/components/parameters/DescriptionFilter'
      responses:
        '200':
          description: >-
            One page of transaction types, together with the two cursor
            positions and the forward availability flag needed to walk the set.
            An empty page is a success rather than a 404: a collection with no
            matching row exists and is empty, and the baseline reports the same
            condition on its own list screen rather than as a failure.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/XCorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionTypePage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - admin-reference-data
      operationId: createTransactionType
      x-required-authority: carddemo-admin
      summary: Create a transaction type
      description: >-
        Creates one transaction type from a supplied code and description. This
        is the migrated form of the add path of
        app/app-transaction-type-db2/cbl/COTRTUPC.cbl, whose screen prompts for
        the new details at its line 158 and asks for a confirming key press at
        line 150 before writing.


        The code is supplied by the caller rather than generated, because it is
        a meaningful two-digit business code that appears in stored transaction
        records and in the concatenated category key, not a surrogate. Creating
        a code that already exists is refused with 409 rather than silently
        replacing the existing row: the primary key declared by
        db/migration/V1__reference.sql raises SQLSTATE 23505, which the shared
        advice reaches through the same integrity branch as a restricted delete
        and therefore answers with the referential sentence the Conflict response
        gives. The divergence that follows from sharing one branch is registered
        as D-REFERENCE-INTEGRITY-SENTENCE in
        docs/architecture/cobol-to-service-traceability.md.
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
      requestBody:
        required: true
        description: >-
          The code and description of the type to create. Both are required:
          the description column is declared NOT NULL by
          db/migration/V1__reference.sql, so there is no state in which a type
          exists without one.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionTypeCreateRequest'
      responses:
        '201':
          description: >-
            The type was created, and the stored representation is returned so
            that a caller sees the row as it now stands rather than assuming its
            own request body was stored unchanged.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/XCorrelationId'
            Location:
              description: >-
                The path of the created type, formed from the code supplied in
                the request body.
              required: true
              schema:
                type: string
                examples:
                  - /api/v1/reference/transaction-types/08
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionType'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/reference/transaction-types/{typeCd}:

    description: >-
      One transaction type, addressed by its own two-digit code. The code is the
      primary key of reference.transaction_types and appears in the path
      directly, because it is a business code a user reads off a screen and
      types back in rather than an internal identifier that has to be concealed.

    parameters:
      - $ref: '#/components/parameters/TypeCdPath'

    get:
      tags:
        - reference-data
      operationId: getTransactionType
      x-required-authority: authenticated
      summary: Read one transaction type
      description: >-
        Returns the transaction type stored under the supplied code, together
        with the version token a subsequent replace must carry. This is the
        migrated form of the single-record fetch
        app/app-transaction-type-db2/cbl/COTRTUPC.cbl performs before it
        displays a record for maintenance.
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
      responses:
        '200':
          description: >-
            The stored transaction type, including its version token.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/XCorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionType'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - admin-reference-data
      operationId: replaceTransactionType
      x-required-authority: carddemo-admin
      summary: Replace the description of one transaction type
      description: >-
        Replaces the description of an existing type. The code itself is not
        replaceable: it is the primary key and the value the category key is
        built from, so changing it would orphan every category beneath it. A
        caller that needs a different code creates a new type and deletes the
        old one, which is exactly what the baseline requires of a user too.


        The request must carry the version token read with the record. When the
        stored row has moved on since that read the write is refused with 409
        and the baseline's own wording, which is the migrated form of the
        before-image comparison the maintenance screen performs across the gap
        between screen turns -- see the Conflict response.
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
      requestBody:
        required: true
        description: >-
          The replacement description together with the version token read with
          the record.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionTypeReplaceRequest'
      responses:
        '200':
          description: >-
            The type was replaced, and the stored representation is returned
            carrying its new version token, so that a caller can issue a further
            replace without re-reading.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/XCorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionType'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - admin-reference-data
      operationId: deleteTransactionType
      x-required-authority: carddemo-admin
      summary: Delete one transaction type, unless categories still reference it
      description: >-
        Deletes the type stored under the supplied code. The delete is refused
        with 409, and no row is removed, while any transaction category still
        references the type. That refusal is the single load-bearing rule of
        this context and it is not an incidental integrity error -- see the
        Conflict response, which carries the baseline's own wording for it.
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
      responses:
        '204':
          description: >-
            The type was deleted. No body is returned, because the resource the
            caller addressed no longer exists and there is nothing to represent.
            The baseline reports the same outcome on its screens with the
            wording recorded on the ReferenceMessageCatalogue schema.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/XCorrelationId'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        A JSON Web Token issued by the configured identity provider and validated
        as a resource-server token. Its cognito:groups claim is converted to
        authorities by
        services/common-lib/src/main/java/com/carddemo/common/security/JwtRoleConverter.java,
        where the baseline user type 'A' becomes carddemo-admin and 'U' becomes
        carddemo-user. Presenting no usable token yields 401; presenting a valid
        token that lacks the authority an operation requires yields 403.
  parameters:
    TypeCdPath:
      name: typeCd
      in: path
      required: true
      description: >-
        The two-digit transaction type code identifying the type, or the parent
        type of the category, being addressed. Seven codes are seeded, '01'
        through '07'.
      schema:
        $ref: '#/components/schemas/TransactionTypeCode'
...
    Cursor:
      name: cursor
      in: query
      required: false
      description: >-
        The paging position to continue from: copy the lastKey member of the page
        in hand to advance past it, or the firstKey member to retreat from it, and
        state the matching direction. Omit it, and the direction, to obtain the
        opening page. The value is opaque and must be echoed exactly as received;
        it is not a type code, a category code or a row ordinal, and it cannot be
        constructed, parsed or compared by a client. A cursor that cannot be
        opened, or one replayed against a different filter, a different collection
        or the direction it was not minted for, is reported as 400 rather than
        answered with the wrong page.
      schema:
        $ref: '#/components/schemas/CursorToken'
    PagingDirection:
      name: direction
      in: query
      required: false
      description: >-
        Which way to move from the supplied cursor. It is meaningful only
        alongside a cursor, and the two are sent together or not at all. Reading
        next takes rows from the cursor's position forward in ascending key order;
        reading previous takes rows before it in descending key order and presents
        them ascending, which is the migrated form of the backward cursor at
        app/app-transaction-type-db2/cbl/COTRTLIC.cbl lines 354 to 368. The two
        differ in whether the cursor's own position is included, as recorded on
        the cursor parameter.
      schema:
        $ref: '#/components/schemas/PageDirection'
    TypeCodeFilter:
      name: typeCode
      in: query
      required: false
      description: >-
        Narrows the page to the single transaction type whose code equals this
        value. This is the migrated form of the optional equality predicate the
        baseline carries in both cursors, at
        app/app-transaction-type-db2/cbl/COTRTLIC.cbl lines 344 to 345 going
        forward and 360 to 361 going back, each gated on an edit flag whose
        unset state passes every row through -- the pass-through arms at lines 346
        and 362. Omitting this parameter is that unset state, and so is sending
        '00': the baseline collapses low values, spaces and zeros into one
        filter-blank state at lines 1101 to 1106. That is why this parameter takes
        TransactionTypeCodeFilter rather than the stored-code schema, which refuses
        '00' outright. A value that is not two digits is reported as 400 carrying the
        baseline's own wording, which is recorded on the ReferenceMessageCatalogue
        schema.
      schema:
        $ref: '#/components/schemas/TransactionTypeCodeFilter'
    DescriptionFilter:
      name: description
      in: query
      required: false
      description: >-
        Narrows the page to rows whose description CONTAINS this text. The baseline
        semantic is a SQL LIKE, at app/app-transaction-type-db2/cbl/COTRTLIC.cbl
        lines 347 to 349 going forward and 363 to 365 going back, gated on a second
        edit flag with the same pass-through arrangement at lines 350 and 366 -- but
        the wildcards are supplied by the SERVICE and not by the caller: the baseline
        strings a percent sign before and after the trimmed input at lines 1155 to
        1163 before binding it. A value with no wildcard of its own therefore matches
        any description containing it, not only one equal to it. Send the text being
        searched for, not a pattern. This parameter takes ReferenceDescriptionFilter
        rather than the stored-description schema, because the baseline applies none
        of its maintenance-screen character rules to a filter.
      schema:
        $ref: '#/components/schemas/ReferenceDescriptionFilter'
...
    CorrelationIdHeader:
      name: X-Correlation-Id
      in: header
      required: false
      description: >-
        Identity of the unit of work this request belongs to, propagated into the
        service log record and echoed on the response. Supply a value to tie this
        request to work already in progress upstream; omit it and the service
        mints one. A value that does not fit the published bound and character set
        is refused with 400 before any handler runs. A value whose shape is that of a
        primary account number -- only digits and the three separators, with
        thirteen or more digits -- is refused on the same terms, because this
        identity is echoed and logged. The refusal carries the shared problem body
        and the response correlation header, the same as any other 400, and it names
        the rule it applied without ever repeating the value it refused.
      schema:
        $ref: '#/components/schemas/CorrelationId'
  headers:
    XCorrelationId:
      description: >-
        Identity of the unit of work that produced this response. Echoed from the
        request where the caller supplied a conforming value, and minted by the
        service where it did not. Quote it when reporting a failure, because it is
        the route from this response to the operational record.
      required: true
      schema:
        $ref: '#/components/schemas/CorrelationId'

    # WHY : Refactoring Rationale: this header is declared because the service
    #       actually sends it and an earlier revision of this document did not say
    #       so. The 401 is rendered by the shared handler
    #       common/error/ApiErrorSecurityHandlers.UnauthenticatedEntryPoint, which
    #       delegates to the framework's BearerTokenAuthenticationEntryPoint FIRST
    #       and only then writes the body, precisely so this challenge is composed
    #       exactly as the OAuth 2.0 bearer-token specification requires. A client's
    #       token-refresh logic reads this header, so leaving it undeclared left a
    #       generated client with no typed access to the one header it needs on the
    #       one response it needs it on.
    # WHY : Assumptions: the value is not enumerated and no pattern is written for
    #       it. The framework composes it from the error it detected -- a bare
    #       scheme when no token was presented, and a scheme with an error code and
    #       description when one was presented and rejected -- so pinning a shape
    #       here would either be wrong for one of those two cases or would have to
    #       be widened until it constrained nothing.
...
  responses:
    BadRequest:
      description: >-
        The request could not be accepted as supplied. This covers a malformed or
        out-of-range path or query value, a body that fails a declared constraint,
        a cursor that cannot be opened or was replayed against the wrong filter,
        collection or direction, a correlation header outside the published bound,
        and a write whose submitted values are identical to the ones already stored.
        The fieldErrors array names every offending field rather than only the first,
        for the reason recorded on the FieldError schema.


        The unchanged-submission refusal is the one member of that list that is a
        business rule rather than a syntactic check, and it belongs here rather than
        on 409 because it is decided from the request alone, before any row is
        contended: the baseline compares the submitted values with the ones it
        fetched and re-displays without writing. Its sentence is carried verbatim and
        is published on the ReferenceMessageCatalogue schema in both of the two
        wordings the baseline declares -- noChangeDetectedSingleRecord for a
        single-record replace and noChangeDetectedList for the list screen -- with a
        field entry naming the member that was resubmitted unchanged.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/XCorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

    # WHY : Refactoring Rationale: this response now states WHICH code and sentence
    #       the body carries, because an earlier revision promised the shared problem
    #       shape without naming its contents and the security chain did not produce
    #       one at all. A refusal decided by the filter chain never reaches a
    #       controller, so it never reaches the shared @RestControllerAdvice, and the
    #       framework default answered with a status and a challenge header and no
    #       body -- the exact opposite of what this response declared. The chain now
    #       installs common/error/ApiErrorSecurityHandlers, whose 401 renderer emits
    #       CARDDEMO-0401 with the sentence below.
    # WHY : Assumptions: the correlation identifier in the body is the SAME value as
    #       the header beside it and as the log line for this refusal, the renderer
    #       reading it from the diagnostic context the correlation filter established
    #       ahead of the security chain. That equality is the whole reason the header
    #       is required on a response whose body names no reason.
    Unauthorized:
      description: >-
        No usable bearer token was presented: the header was absent, the token was
        malformed, its signature or issuer did not validate, or it had expired.
        This is distinct from 403 and from 409, and the distinction is not
        cosmetic -- a caller acts on 401 by obtaining a token, on 403 by obtaining
        a different one, and on 409 by changing its request. No body member names
        the group the caller lacks, because at this point no verified identity
        exists to compare against one. The body carries code CARDDEMO-0401 and the
        sentence "Sign on to continue", identically on all four of those causes,
        alongside the bearer-token challenge below.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/XCorrelationId'
        WWW-Authenticate:
          $ref: '#/components/headers/WwwAuthenticate'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
...
    TransactionTypePage:
      type: object
      description: >-
        One page of transaction types together with everything needed to walk the
        ordered set by key in either direction. When the filters match no row the
        page is empty and both cursors are null; the baseline reports that
        condition with two different sentences depending on which of its screens a
        user was on, both published on the ReferenceMessageCatalogue schema.
      required:
        - items
        - firstKey
        - lastKey
        - hasNext
      additionalProperties: false
      properties:
        items:
          type: array
          description: >-
            The transaction types of this page, in ascending type-code order.
            Empty, never null, when the set is exhausted or the filters matched
            nothing. Its length must not be used to infer the end of the set --
            read hasNext for that.
          items:
            $ref: '#/components/schemas/TransactionType'
        firstKey:
          oneOf:
            - $ref: '#/components/schemas/CursorToken'
            - type: 'null'
          description: >-
            Names the first row actually returned in items, or null exactly when
            items is empty. It is also the position a caller pages backward from:
            send it as the cursor with direction previous. Null here means there is
            no backward position to send, which is the one case a caller handles
            rather than retries.
        lastKey:
          oneOf:
            - $ref: '#/components/schemas/CursorToken'
            - type: 'null'
          description: >-
            Names the last row actually returned in items -- never the extra row
            read to settle hasNext -- or null exactly when items is empty. It is
            also the position a caller pages forward from: send it as the cursor
            with direction next.
        hasNext:
          type: boolean
          description: >-
            Whether a further page follows this one, settled by reading one row
            beyond the page and observing whether that row exists, never from any
            tally. When it is true, lastKey is present. When it is false, a caller
            that presses forward anyway receives the baseline's own end-of-set
            wording, published on the ReferenceMessageCatalogue schema.
      examples:
        - items:
            - typeCd: '01'
              description: Purchase
              version: 1
            - typeCd: '02'
              description: Payment
              version: 1
          firstKey: v1.dHJhbnNhY3Rpb24tdHlwZXM.REDACTED-EXAMPLE-SIGNATURE-not-a-real-mac-F
          lastKey: v1.dHJhbnNhY3Rpb24tdHlwZXM.REDACTED-EXAMPLE-SIGNATURE-not-a-real-mac-L
          hasNext: true
...